backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pymongo>=4.6.1
motor>=3.3.2
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.0.0
pdfplumber>=0.10.4
python-docx>=1.1.0
firebase-admin>=6.4.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
jinja2>=3.1.3
PyPDF2>=3.0.1
google-generativeai>=0.8.0
razorpay>=1.4.2
